Tools/memory_v2/requirements.txt
# Thanos Memory V2 Dependencies
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
mem0ai>=0.1.0
voyageai>=0.3.0
numpy>=1.24.0
